Make RandomObjectPopulator background persist thread a daemon thread to simplify shutdown logic due to guice creation errors
Since the guice injector is created with stage PRODUCTION and almost all
providers are marked as singleton, a situation arises where a CreationException
is thrown while creating the injector, but after the object manager is created.
This makes it impossible to get a reference to the object manager in order to
properly shut it down, and the tool will hang. The risk of making this thread
daemon is that exposes the risk that the caller will not call shutdown properly,
so objects could be lost.