Rework Client handling of Content-Length and Transfer-Encoding headers
The Apache HttpClient library will set the appropriate headers based on
the configured RequestEntity for the request that is being executed. If
the caller has already explicitly set either Content-Length or
Transfer-Encoding, HttpClient will throw a ProtocolException.

Previously, OG's Client implementation would explicitly remove these
headers in advance in the case of a redirected request. Now, because
some auth implementations may add one of these headers as a part of their
authentication algorithm, it is required that the Client implementation
always explicitly remove these headers prior to execution of any request.