Remove getMessageTime from Message interface
This method was used as a workaround because the existing HttpRequest
implementation automatically inserted a Date header at construction
time and it was required to be able to override this header for
authentication testing.

With the removal of this interface method, the HttpRequest
implementation has also been changed to not automatically insert a Date
header when constructed.