Add AWSV4 auth digest caching
The message digest computation for AWSV4 auth is CPU intensive. In order
to alleviate the cost of doing this repeatedly for both standard and
chunked content signing, add a message digest cache. In order for the
cache to be effective, the source data must be configured to use all
zeroes rather than semi random data.