Rework RequestSupplier to use context Functions
Previously, Function implementations optionally would either add entries
to the context map or read entries from it. This distinction was not
made clear and the ordering of the functions was brittle depending on
the ordering. Now, a list of functions with a determined ordering are
called first to populate the map, and then every other function in
RequestSupplier may optionally read from it.